requirements.txt
streamlit
phidata
openai
pdfplumber
python-docx
google-generativeai
duckduckgo-search